Revert HybridWebView refactoring changes from PR 32491 to address regression issues
