+++ src/Core/src/Handlers/HybridWebView/HybridWebViewHandler.Tizen.cs
@@ -8,8 +8,6 @@ namespace Microsoft.Maui.Handlers
 
 		public static void MapEvaluateJavaScriptAsync(IHybridWebViewHandler handler, IHybridWebView hybridWebView, object? arg) { }
 
-		public static void MapInvokeJavaScriptAsync(IHybridWebViewHandler handler, IHybridWebView hybridWebView, object? arg) { }
-
 		public static void MapSendRawMessage(IHybridWebViewHandler handler, IHybridWebView hybridWebView, object? arg) { }
 	}
 }
